+++ V-Wallet_backend/Application/Commands/CashOutCrypto.cs
@@ -7,5 +7,5 @@ public class CashOutCrypto : IRequest<CryptoCurrency>
 {
     public string UserId { get; set; }
     public long CryptoId { get; set; }
-    public long CurrentValue { get; set; }
+    public decimal CurrentValue { get; set; }
 }
