+++ V-Wallet_backend/Application/Commands/CreateCryptoInvestment.cs
@@ -7,6 +7,6 @@ public class CreateCryptoInvestment : IRequest<CryptoCurrency>
 {
     public string UserId { get; set; }
     public string Name { get; set; }
-    public long Value { get; set; }
-    public long Investment { get; set; }
+    public decimal Value { get; set; }
+    public decimal Investment { get; set; }
 }
